Add clickedOnLine state and setClickedOnLine prop***
This commit adds a new state variable called clickedOnLine and a corresponding prop setClickedOnLine to the CustomMapEvents component. This state variable is used to track whether a line feature was clicked on the map. The setClickedOnLine prop is used to update the value of clickedOnLine. This change allows for conditional logic in the click event handler to prevent executing certain code when a line feature is clicked.